Do some crazy stuff to make `solve_one` work
Lot of experimental stuff with overriding class methods and attribute
accesses. The problem here is that `solve_one()` creates its own
`Con_solver`s, but we lose a reference to them (we only have the one we
originally created). The only way around that is to override
`Con_solver` methods directly. This is getting really hacky...