Refactor how nodes/arcs are identified from Python to JS
We now save IDs of the generated JSON graph in Python, such that we can
map from variables and constraints to the corresponding ID. This way,
our graph visualizer has to simply find the corresponding ID and go
along its way, without needing to know about constraint names and such.

There was also renaming of fields and more documentation to make things
clearer.