Enable multiple views of the same CSPViewer to be shown
By registering event handlers for the Python code into the model as a
one-time operation, instead of every time a view is created, this
prevents issues where the various CSPViewer's (of the same instance)
become out of sync.

This adds a bit of complexity to the code, mainly that events have to be
forwarded from the model to the views after being received from the
Python code. Also, because views may not be created when the model
(both the JS and Python one) are created, we need to flag when the
initial rendering is so we can apply the initial state (make all arcs
blue). All further renderings of the model sync to the current state.