Implement a run loop that is ran on a separate thread
Instead of running each function (`make_arc_consistent` and `solve_one`)
on a separate thread each time their corresponding button is clicked, we
run one function that handles switching between these two. The benefits
is that we retain one thread for each visualization only, instead of
spawning multiple threads and throwing them away. We expect using one
thread to also help with identifying the "main" `Con_solver` as well.

There is also an experimental "kill switch" to make this work. By
throwing an exception in the display method (though an exception should
probably be thrown in any user-input method), we can use that to signal
the current method execution should end. This isn't very robust - it
will not be able to jump out of a `while True` loop, for example, if it
doesn't call `display` - but there doesn't seem to be another way.