Move controls to front-end
Using ipywidgets in the backend was convenient, as the functions that
needed to be called were right there. However, there isn't a way to
control where those controls ended up (resulting in them to always be
placed at the top) and they weren't part of the same container as our
visualization. Because of that, we create the buttons in HTML instead.
The added complexity is that we now need to send messages for our
buttons to communicate with Python! This is simple enough, but results
in quite a bit of boilerplate code, unfortunately.