Pass in visualizer directly to `Con_solver`
To solve the problem of giving our `Con_solver`s references to the
Jupyter widget displaying them, we've previously resorted to
monkey-patching methods that are bound to the Jupyter widget at run
time. While this worked, it is quite a brittle approach, plus if
additional `Con_solver`s were created within the original `Con_solver`,
we have no way of patching those (patching the class's methods doesn't
work when there are two cells each creating a `CSPViewer` - the last one
created would change all the patched methods to point to it). After a
meeting, we've decided that we were okay with the tradeoff of adding a
visualizer argument to every `Displayable`, giving us a direct reference
to our Jupyter widget as desired. This little addition allows us to
remove all the monkey-patch hacks we'd resorted to, at the cost of a
little more code presented to the user that isn't algorithm-related.