Allow directly creating a `Con_solver` and queuing methods on it
One problem we found is that there is fundamentally a disconnect between
the wrapper Jupyter widget and `Con_solver`. There is no explicit
indication when creating a `CSPViewer` that it creates a `Con_solver`.
Furthermore, there is also a disconnect between which function you are
calling whenever anything happens.

The solution to the former issue is directly creating a `Con_solver` and
passing it directly to the `CSPViewer`: that establishes the connection,
as desired. Another solution is to make `Con_solver ` a Jupyter widget,
such that it can be created and displayed directly. That is the approach
taken here. One drawback is that all code now must be hidden inside
`Displayable` to avoid polluting `Con_solver`'s user-facing code; that
makes `Displayable` unique to `Con_solver` and not reusable. Perhaps we
could remedy that by importing a different `Displayable` for each class
that can be displayed.

The latter issue is a bit more thorny. If we naively call the functions
directly, due to the presence of blocking functions (to wait for user
input, for example), then the function will not complete and nothing is
ever displayed on the screen! Calling it after also causes it's own set
of problems (and also very annoying - you need a separate cell to do
it!). The approach here is to introduce a decorator that queues the
function call, and executes it once displayed. Further calls of that
method (e.g. recursively) do not queue the method. What's particularly
great about being able to do this is that it partially solves the problem of
 what happens when a new method is added by the user (for example, adding
`solve_all` as part of an assignment). How do they call that function?
Previously, there was no direct way to do so. We still need to consider
how to allow the user to add new buttons or further control the
visualization, though.

This allows us to call `Con_solver()` and
`Con_solver().make_arc_consistent()` and have things display properly.
As always, the tradeoff is more code presented to the user, and perhaps
unintutitive behaviour (regarding queued function calls).