Create separate Displayable to display CSPs explicitly
The current thinking is that explicit is better than implicit. If we
choose to make `Con_solver` a Jupyter widget (which necessitates
`Displayable` to become tied to it), then we would either need different
`Displayable`s for each class or have some logic inside one
`Displayable` that switches between different implementations, depending
on the class being instantitated.

What's nice about using separate `Displayable`s is that it paves the way
to allowing users to create classes that display, e.g. CSPs, without
needing to modify the existing machinery (`Displayable` is intended to
be an implementation detail). As long as they call the right display
methods (which will probably need some changes sooner or later, as
string matching isn't fun), it should "just work".

For console users, the change is mainly a different import and an
annoying decorator (`@delay_until_render`) to make everything work.