Test implementation of splitting a clicked variable
When running `solve_one` and after arc consistency completes, you can
click on a node to split. This will split the node in half if it has
more than 1 element. However, the visualization doesn't reflect this
because we don't actually know `solve_one` is being called again with
different domains. Previously we handled this by monkey-patching
`solve_one`, triggering domain updates every call, but since we are
moving away from that, passing in the visualizer explicitly instead,
that doesn't seem like an enticing option.