Begin working on graph creation; enable deleting nodes
The current idea is to have a specialized Jupyter widget that will spawn
a graph-building cell. Then, the resulting CSP can be fetched from it
in another cell, e.g. assuming `c` is an instance of `CSPBuilder`,
`c.get_csp()`.

While implementing the node deletion functionality, it was clear that
some refactoring had to be done to the visualizer to better separate
concerns with regards to one-time setup events and re-rendering draw
code.