Test converting created CSP into Python object/code
Our GraphVisualizer uses JSON to represent a graph, so in order to
render a Python CSP, we first have to turn it into a JSON and pass it to
JavaScript.

The same applies in the other direction. Our CSPBuilder internally
tracks a JSON object to represent the graph, updating it as it goes
along. However, we need to get this into a Python CSP object such that
it can be rendered by CSPViewer (and so functions like
`make_arc_consistent` work). So, we need to convert JSON back to CSP. We
can retrieve the CSP with `csp()`. We are also testing a `py_code()`
function on the CSPBuilder that replaces the current cell with the
corresponding Python code that can be executed to create the CSP.

One major issue is how do we represent constraints? Since arbitrary
Python code can be used to write a constraint, how can we serialize it
into JSON format? We could of course try to stringify the code and
`exec` it later, but this may cause security implications.