Complete revamp of GraphVisualizer internals
Initially, the graph representation was a JavaScript object that
closely resembled what d3's force layout expected, to make it easy to
use it for layout purposes. This became annoying though, because force
layout would add properties to the nodes, breaking types, and also made
things like drag support more complicated (you had to restart the
"simulation", despite fixing all the node's positions after the first
layout).

The decision was made to break away from this representation. The
benefits are two-fold: first, we didn't have to tie our implementation
to how force layout worked, opening the doors to the idea of swappable
layout engines. Second, transitioning from a simple JS object to a class
lets us use powerful methods to manipulate our graph, abstracting away
the internal representation as much as possible.

The result is changes to pretty much the entire codebase. There is still
some work to be done to expose less of the internal node/edge
representation and to nail down a JSON interchange format.