Add additional display call to synchronize domains
Before, domains did not update when `solve_one` was being called
recursively, because even though we knew that the domains had been
split, we didn't know which "branch" of the recursive call we were at.
Our visualization had no idea that `solve_one` was being called again
with a new set of domains, and so kept updating the original set of
domains it knew. This made the visualization out of sync with what was
actually happening in code.

We got around this issue before by monkey-patching `solve_one`, such
that every time it was called, we would send a domain synchronization
message to the front-end. Since we are moving away from monkey-patching
due to brittleness, we no longer can "listen" for a certain function to
be called. Our solution instead is to explicitly add a display method
that alerts us of the domain changes.