Replace Graph builder with Vue-based component
This is slightly crazy, but hear me out: I've decided to try out using
Vue.js for all my UI components instead of D3.js and framework-less
JavaScript. The reasoning for this starts when I was attempting to
implement a little toolbar where I could edit the currently selected
node's name and domain. This quickly got crazy (see 80b5a7921),
devolving into manually syncing UI with state and so on.

Another reason was just that the existing hierarchy of inheritance was
quickly devolving. I wanted a way to make a generic GraphBuilder that
provided base functionality, like creating nodes, available, and allow
overriding to configure what nodes are created. However, I had
GraphBuilder inheriting from CSPGraphVisualizer because I wanted to
piggyback off the node rendering (the right shapes and contents). That's
no good. I couldn't see a clean way to "swap in" node rendering while
still handling the differences between a graph builder (say, selecting
nodes and arcs) vs. the graph interactor (changing node colour on mouse
over, etc). I also was being to grow frustrated with D3. I wasn't using
it's power, mainly just as a glorified DOM API, and too many rendering
details (like binding to data on a specific selection) was leaking to
subclasses.

That lead me back to a component-based approach, which I had grown to
love. After considering React again, I saw Vue as a balance between
React and Angular with a very approachable API and a vibrant community.
Because of it's emphasis on one-way data flow, it makes synchronizing UI
to state trivial. Also, I found that by using slots, it made swapping in
and out nodes possible (though there is still much room for
improvement).

The last thing was the decision to use Vue with Backbone. Backbone is
the framework Jupyter widgets used internally. So it is definitely weird
to slap another framework on top (albeit a UI framework). The reason why
I see this as being okay is that a) Backbone isn't opinionated with what
you use to render and b) JupyterLab isn't using Backbone anyways, so why
tie myself to it? Instead, I could choose whatever I wanted and port
that around.

Hopefully this choice pays off. I had a lot of fun working with Vue, and
it has made things significantly easier UI-wise. This is, as always, a
WIP.