Merge Graph (Builder/Interactor) into common base
There use to be GraphBuilder and GraphInteractor, both of which had to
set up the SVG and thus duplicated a lot of common knowledge. However,
the only differences between them ended up being what events were being
exposed. Since these are rather useful in any situation, we merged them
into one base class.

Using the power of named and scoped slots, we now have a way of
powerfully exposing the differences in each type - like shape of nodes
and their effects when clicked or hovered - while still wrapping some of
the setup. We have chose to not have too many "base functionality" - for
example, even the concept of a node/edge being selected isn't built-in
to the base class. The reason for this is that not all types have the
concept of selection, and there are times when you'd like to reject
or change selections in response. Instead, we just expose the click
event and have the parent component manage it. This duplicates a bit of
code if you always need selection, but this is the cost of flexibility.