Revert synchronization method introduced in 456fb665
Unfortunately, this approach doesn't work with `builder.csp()`, which is
synchronous. I played around with the IPython event loop, but ticking it
forward in a busy-wait loop unfortunately dissociates the output from
the executing cell, instead moving it to where the variable was
declared (see https://github.com/jupyter-widgets/ipywidgets/issues/1188
for more details). I guess because I only need to do these updates for
the builder, and not the interactor, it's "okay" that I need to take
this performance hit of synchronizing it all the time.