Reformat front-end files with Prettier
Prettier replaces tsfmt as the formatter of choice, given that it
supports JS, TS, and CSS. Instead of having a lot of different
formatters for each file type, we stick with one. It follows the same
thinking of having a consistent rule-base. There was also some issues
with tsfmt formatting in a way that tslint didn't like - we have added a
set of rules that prevent tslint from commenting about formatting, using
it just as a linter.