Show edge costs on DirectedEdge
DirectedEdge can now show a horizontal label in the center of the edge
(there are still edge cases, e.g. we should probably remove the label if
there is no room to show it). This was surprisingly not as simple as I'd
hope, because you can't actually have a text background like you can in
CSS! You need to draw a rect yourself.

The other thing I realized is that the width/heights of everything
weren't available in Vue because at the time of render, the el in
render() wasn't available. The solution is to wait a tick, so the DOM is
initialized, such that when we add our Vue instance, it has a sized
parent.