Highlight nodes and paths during search visualization
All Search nodes now have a stroke and strokeWidth property that is used
to do the highlighting. The colours and what is highlighted during the
visualization is not final, but closely mirrors what is available in the
applet so far. The exception is that there is no support for the
"added to frontier" event, as the existing Python code only reports the
entire frontier, not what is added.