Run searcher's search method on a background thread
The Searcher is run similarly as it is in the terminal - you repeatedly
call search on a Searcher instance to advance the internal state. Unlike
CSP's visualize, we don't need to queue a function to run, because it
always runs in a background thread, whether or not it is the first
render or not (in CSP's, we had to do this because it was recursive -
we didn't want to run the recursive function in it's own thread!).

One caveat is that if you try running search() again before it finishes
- since it has popped off something from the frontier, but has not yet
had a chance to add anything back, you get "no solutions".