Convert DirectedEdge to use TypeScript
I've been thinking about using vue-class-component to make things even
more TypeScript-y, but I'm not sure how stable decorators are, and my
attempts with it made VSCode very angry at me for using decorators
despite setting the appropriate options in tsconfig.json. Anyhow, my
main purpose for this is a) to get type checking while using `this`, and
b) a place to document everything. Tooling support definitely is a
weakness of Vue, but I'm optimistic it will improve.