Insert generated code below cell instead of replacing it
Replacing the cell breaks if you use 'run cell, select below', and this
was a problem when running the entire notebook at once. Adding a new
cell every time sidesteps this issue, but it may get annoying if you end
up with a lot of "versions" of your graph, and have to manually delete
them.