Update node arrow to be the same colour as the line
SVG markers don't support different colours(!), which make it very
annoying to deal with (you need to create, on the fly, markers for each
colour). Instead, we just use a polygon and orient it to the end of the
path ourselves, giving us full control over the colour and the size of
the arrow.