Search nodes now can dynamically scale based on text
Support for this touches two areas: first, the search node must be able
to calculate it's own width and height, which it does with some simple,
hardcoded calculations (i.e. doesn't actually try to measure text).
Secondly, the DirectedEdge must know about this new sizing so that it
can draw its edge properly.