Add automatic relayout when window size changes
Currently implemented in both visualizers (not creators, because we need
to tackle recalculating the layout the user has dragged themselves).
This commit removes the limitation of a fixed graph size, and instead
adopts whatever size the output area in the Jupyter notebook is. It does
this by listening for resize events, and after a delay to prevent too
much computation, lays out the graph again. There is more work to be
done to make this nicer (we don't actually use relayout, and perhaps all
of this code should be inside the graph component itself?), but it works
well in practice.