Support text truncation on search nodes
A truncation algorithm has been added to search nodes that scale the
node width based on the text up to a certain maximum, then truncates it.
We actually render the text and measure the change, recursively removing
characters until we are under the desired width. This is usually quite
fast because we already have an idea of where text should be truncation
(the previous truncation length), and only need to take into account any
newly added characters, to see if they affect what should/should not be
truncated.