Unified search nodes into one object; support for subtext
A new component, EllipseGraphNode, combines all the behaviour of the
previous search nodes together into one. This node is intended to be
used beyond search. It supports auto-resizing and text truncation, as
before, but also supports a second line of text, the subtext, with
auto-truncation and sizing based on that as well.