Migrate all Vue components to Typescript classes
At the cost of a little extra complexity, all Vue components now are
Typescript classes thanks to the use of decorators. This makes the
classes fully type-checked, meaning great Intellisense. The previous way
of annotating the Vue components with interfaces caused a lot of
duplication and a disconnect between what was props/data/computed. I
also took the chance to run Prettier on all the components, as supported
by prettier-atom.