Remove object-literal-sort-keys rule from TSLint
This was quite possibly one of the most annoying rules enforced in the
recommended configuration, and a quick GitHub search shows that this
rule is consistently removed. It got in the way when you want to specify
objects with common properties (like name and id) first, then put the
custom ones at the end, and didn't allowing grouping keys (like width
and height) together.