Add dynamic graph generation from implicit search problem
The restriction that a search problem had to be explicit to be
visualized has been lifted. This is useful for (and created to support)
solving CSPs as search. The graph is generated on the fly as neighbours
are "discovered".

This required a lot of changes:
- A tree layout has been added. The search space for solving CSPs with
search is a tree, and force layout did not generate anything close to
a tree. For search only right now, the `layout_method` can be switched
from "force" to "tree" to enable this behaviour. Unfortunately, this
didn't work "out of the box" because the expected data structures
between force layout and tree layout are very different.
- Comm messages have been consolidated. A problem found when testing the
performance of solving CSPs was that sending many small Comm messages
(from the backend to the frontend) was very slow. Work is being done to
send larger, but fewer, messages. An example of this is highlighting
nodes - instead of sending a message for each node to be highlighted, we
now send an array of nodes to be highlighted (if their colours are the
same). In the future, we could possibly just send one message every
`display()`.
- The "Neighbours" display call had to be overhauled, to handle the
implicit case. If it detects the search problem is implicit, it will add
the new arcs discovered if not already there to an explicit
representation. It is this explicit representation that is ultimately
visualized, allowing for reuse of existing front-end code.
- Nodes stored in the JSON representation are now explicitly casted to a
string. This was an issue because nodes in CSPs are represented as
dictionaries, which are not usable as dictionary keys. `str()` is usable
right now, but may cause conflicts later on by not being unique enough.
`__repr__()` is another option.