Move Jupyter implementations around
We have decided to create two copies of each algorithm: one, in the
aipython module, will be identical to those in the book - running them
will output text, as if you were running them in the terminal. The
second copy will be in the aispace2 module, which is where the (possibly
interactive) algorithms and their visualizations exist. This prepares
for some of this work, by moving all visualization backends into the
jupyter folder, in order to be explicit.