Make IDs persistent across one run
Node and edge IDs were randomly generated, which was a problem when the graph was converted to and from JSON, making even the same nodes have different identifiers. By using their hash instead, the IDs are unique and static across one run of the program. This allows us to identify the same nodes across graphs, and reapply their styles.