Generate ID -> node/edge mapping client-side
Since the graph is stored as an array of nodes and edges, looking up a
node or edge by ID is linear time. While this is generally fast enough,
given that all our graphs have <100 nodes (any more would be hard to
visualize), it was a bit sloppy. Instead, we generate the mapping every
time the graph is created and so get constant time lookups.