Replace CSP graph nodes with generic ellipse/rectangle nodes
By using generic nodes, we can standardize features for nodes, like
stroke, text colour, and fill colour, alongside automatic sizing and
truncation. Replacing variable nodes with ellipse nodes was easy; search
already paved the way with ellipse nodes, so they were a drop-in
replacement. For constraint nodes, we had to introduce a rectangular
node, which is basically the ellipse node but draws a rectangle instead.
This implementation is much simpler, because we don't need two line
functionality (yet) and it is a bit easier working with width/height
rather than radii. Ideally, the text truncation functionality would be
pulled out into a mixin or something, so any node shape could take
advantage of it.