Animate property changes of nodes/edges
A new prop has been added for node and edge containers, called 'transition', that
enables transitions on their children.

This is super helpful in dynamically updating graphs (such as solving
search as CSP) as the nodes move to their final destinations. This is
less jarring than snapping them to their final destination right away.
As a stylistic choice, fills and strokes are not animated.