Overhaul development webpack configuration for speed
webpack was being too slow for my liking recently, with initial compile
times around 20 seconds and subsequent builds about 3 seconds. This was
far too long for a relatively small project, so I spent a lot of time
hopefully making it faster.

Development builds now use DLLs. This means library files are bundled
separately, so the files that rarely change do not have to be
re-bundled. This results in faster build times (about 5 seconds faster).
However, the downside is that DLL's must be regenerated when libraries
are updated. This is quite confusing, and can cause problems where you
think you've updated your dependencies but the vendor code hasn't.
Because of that, I've decided to ALWAYS run the DLL task before any
development command. This removes most of the benefit of speed up
during initial buid (it is somewhat faster because it doesn't
need to create as large of a file and can more easily resolve library
dependencies), but retains the speed benefits of
recompilation. I felt this tradeoff was worth making it dramatically
less confusing, but perhaps it is not. Granted, the DLL task only takes
about 2 seconds right now.

Another improvement is switching Typescript to transpile-only. Hold
on... doesn't this defeat the purpose of Typescript? Well, we now use a
webpack plugin to run type-checking on a separate thread, so it doesn't
hurt build times. This makes the build times about 10 seconds faster!

As a result of these changes, build times went from 20s -> 7s for
initial builds, and about 3s -> 0.5s for incremental ones.

Note that there are no changes to production. DLLs aren't conducive for
tree-shaking, so none are used/generated during that process. TypeScript
is also not set to transpileOnly. A slower production build time is
fine, as long as it generates the smallest possible bundle and catches
as much errors as it can.