Simplify embedding of Vue components in Backbone views
We use to use a wrapper Vue component for each actual component, in
order to benefit from the reactivity. However, this means we had to
first configure data in the wrapper Vue component, then pass those same
properties as props to the actual component. This duplicated a lot of
code and just made things unnecessary confusing. Instead, we have
chosen to create the actual component directly, and change all props to
data. This means that it is no longer possible to embed the actual
component inside another template (since it doesn't accept props), but
the tradeoff is that these visualizers are at the "edge" of the
component hierarchy and won't be embedded.