Make Graph visualizers handle layout/resizing
This commit completely revamps how graphs are layed out. Before, the
root component had to do everything in terms of layout - specify the
layout function, listen for window resizing, etc. Now, we've
deduplicated that logic by making the graph visualizer itself handle
these things - the parent now only specifies the layout algorithm. The
width, height, and resizing of the graph is now all handled
automatically. As part of this change, we changed how Vue components
were initialized in Backbone; we are now able to directly pass in the
HTMLElement during mounting, which ensures correct width/height, and
fixed the previous limitation of this approach by correcting the
`destroy` lifecycle method.

Speaking of layout algorithm, there was always the notion of a
`LayoutEngine` that had two functions: `setup`, for the initial graph
layout, and `relayout`, which handled any subsequent changes. This was
never utilized though, as we always just called `setup`, but this has
been fixed so the functions are called as intended. This would allow you
to start the graph off using a force layout, but then only scale
relative positions as a result of resizing (e.g. great for builders).
`LayoutEngine` is now simply `GraphLayout`, which accepts two functions,
one for setup and one for relayout, which components can call as
necessary.