(Re)enable splitting domains by clicking variables
This commit reintroduces the functionality of clicking a variable to split on it, which was available some time before but removed shortly after due to it being finnicky. Also, as part of this change, clicking on vars/arcs only trigger stepping if we expect the user to click on a var/arc - clicking them otherwise will do nothing. Documentation improvements are also provided.