Experimental support for choosing domain to split on
The UI for this is a window prompt, which is really not the right place to go. But it's easy and fast, which has been the Achille's heel of programmers everywhere.

Anyways, what this allows you to do is, after choosing a variable to split on, choose the subset of the domain to search through first. If the user cancels the popup prompt, then the default split (in half) is used.