Add experimental resize handle to SVGs
Implements a resize handle that allows the user to expand or shrink the height of the SVG. This is in addition to the resize that happens when the window is resized (but this only affects the width now, not the height).

CSS has a resize property that would be very helpful here, but it isn't supported on IE/Edge and Safari does not emit any events when it is resized that way (even with a polyfill for ResizeObserver).