Run search function in same cell as search visualization
Instead of having to create the visualizations and then run search,
which was deemed confusing, run search for the user the first time. The
user can run search again after to find more solutions.

This borrows the approach as CSPs, where the function is queued if not
yet displayed. We improved on that implementation by simplifying things
a bit.