Overhaul how JSON <-> Search/CSP works
Initially I used manual synchronization to handle serializaing a search
problem or CSP to JSON - that is, convert it to a dictionary, and sync
that. However, I quickly realized that Jupyter provides a way to do this
for us, by using an Instance traitlet combined with (de)serialization
functions provided as arguments. This is the "correct" way of doing
things, but I put it off for a while since it was just a lot of work for
no added benefit. In the interest of keeping the code as clean as
possible, I have moved to the correct way. The functionality has not
changed (hopefully); but things should be clearer and more direct now.