Graphs in CSP resize whenever we resize the canvas size. This is an issue because it makes it more difficult for users to find the nodes due to their positions may have been changed drastically after each canvas size change.
Relative layout has been applied to prevent the graph from drastically reshape and reposition its nodes after canvas resize. However, due to the new feature of hovering to display full node content, the padding set in relative layout is not sufficient to include all nodes fully in the canvas. So the padding is increased from 30 to 60, where 50 is the maximum size of unhovered node and an additional 10 for a comfortable zone in case of user setting super large font size.