Improved layout of forward planning
This improvement make the graph nodes not swap to a random position at each relayout, it keeps the relative positions unchanged among nodes.

The new layout basically adjusts the left side of the graph.
The nodes that have already been rendered in previous steps will only move slightly to fit the whole graph.

Code for fixed-node layout is included and is commented out.